+++ CHANGELOG.md
@@ -6,7 +6,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 ## [0.7.6] - 2024-07-29
 
-- [shelly]: Refactor mdnsScanner and mdnsScanner jest test.
+- [shelly]: Refactor MdnsScanner and MdnsScanner jest test.
 
 ## [0.7.5] - 2024-07-28